intial changes, trying to use superclass line_detection
